SelfContact setup reimplemented once again.
Description of SelfHandle property of the Connection iface: "If the
connection is not yet in the CONNECTED state, the value of this property
MAY be zero".

Description of SetPresence method of the Presence iface: "Request that
the presence status and status message are published for the connection.

Changes will be indicated by PresencesChanged signals being emitted.

This method may be called on a newly-created connection while it is still
in the DISCONNECTED state, to request that when the connection connects,
it will do so with the selected status."

Argument of PresencesChanged() signal is a map handle<->presence.

Summary: SetPresence called on a newly-created connection leads to
PresencesChanged() signal with self handle (which is legally zero).
In fact, it drives mission-control crazy, which deadly affect some telepathy
client.

To avoid this disco, set selfHandle to 1 on connection creation no matter what.

May be the proper solution is to prevent PresencesChanged() signal from
emission, if handle is zero.