Increase minimum required CMake version to 3.6
As current TelegramQt requires CMake 3.6 or newer, building current
TelepathyMorse against it using a recent CMake (like 3.10) fails with
`Unknown arguments specified` error for `"NOT" "Core" "IN_LIST"
"_WANTED_COMPONENTS"` at `TelegramQt5Config.cmake:10`. CMake warns to
set CMP0057, but doing this in TelepathyMorse does not help as the
minimum required version 3.1 does not know the CMP0057 and TelegramQt
requires a newer version. Increasing the minimum required CMake version
for TelepathyMorse fixes the issue. (I think even increasing it to
version 3.3, that introduced CMP0057, would fix it, but as TelegramQt
requires version 3.6+ it might be better to be the same.)